+++ src/components/KlabSettings.vue
@@ -12,7 +12,21 @@
         @mouseleave.native="mouseFabLeave"
       >
         <q-fab-action
-          v-if="layout !== null || !isLocal"
+          v-if="!isLocal"
+          color="app-background-color"
+          text-color="app-main-color"
+          @click="logout"
+          icon="mdi-power-standby"
+        >
+          <q-tooltip
+            class="klab-app-tooltip"
+            anchor="center right"
+            self="center left"
+            :offset="[8, 0]"
+            :delay="600">{{ $t('label.appsLogout') }}</q-tooltip>
+        </q-fab-action>
+        <q-fab-action
+          v-if="!isApp && layout !== null"
           color="app-background-color"
           text-color="app-main-color"
           @click="exitApp"
@@ -23,7 +37,7 @@
             anchor="center right"
             self="center left"
             :offset="[8, 0]"
-            :delay="600">{{ isApp || !isLocal ? $t('label.appsLogout') : $t('label.appsClose') }}</q-tooltip>
+            :delay="600">{{ $t('label.appsClose') }}</q-tooltip>
         </q-fab-action>
         <q-fab-action
           color="app-background-color"
@@ -149,6 +163,7 @@ export default {
     ]),
     ...mapGetters('view', [
       'isApp',
+      'klabApp',
       'hasShowSettings',
       'layout',
     ]),
@@ -210,6 +225,7 @@ export default {
       });
     },
     exitApp() {
+      /*
       if (!this.isLocal) {
         this.$nextTick(() => {
           if (this.isApp && this.sessionReference.publicApps.length > 1) {
@@ -218,16 +234,19 @@ export default {
             this.logout();
           }
         });
-      } else if (this.layout) {
+      } else
+      */
+      if (this.layout) {
         this.setLayout(null);
       }
     },
     logout() {
+      const url = `${process.env.WS_BASE_URL}${process.env.ENGINE_LOGIN}${this.isApp ? `?app=${this.klabApp}` : ''}`;
       if (this.token !== null) {
         axiosInstance.post(`${process.env.WS_BASE_URL}${process.env.API_LOGOUT}`, {})
           .then(({ status }) => {
             if (status === 205 /* Reset Content */) {
-              window.location = `${process.env.WS_BASE_URL}${process.env.ENGINE_LOGIN}`;
+              window.location = url;
             } else {
               this.$q.notify({
                 message: this.$t('messages.errorLoggingOut'),
@@ -250,7 +269,7 @@ export default {
             console.error(`Error logging out: ${error}`);
           });
       } else {
-        window.location = `${process.env.WS_BASE_URL}${process.env.ENGINE_LOGIN}`;
+        window.location = url;
       }
     },
     mouseActionEnter(actionName) {