+++ src/i18n/en/index.js
@@ -194,7 +194,7 @@ export default {
     noGroupsAssigned: 'No groups assigned',
     appLoading: 'Loading app <strong>{app}</strong>',
     errorLoadingApp: 'Error loading app <strong>{app}</strong>',
-    errorLoggingOut: 'Error logging out, contact with support',
+    errorLoggingOut: 'Error logging out, contact support',
     youHaveGOT: 'Winter is coming',
   },
   tooltips: {