+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project versioning adheres to [Semantic Versioning](https://semver.org/
 is not available.
 
 **The version of k.Explorer software is self-referred and is not linked to the k.LAB version, so is possible that some features are not available in the stable version of the engine.**
+## ?
+### Features
+- Rebrand
+### Features
+- Logout and exit app management for remote engine
+
 ## 0.12.1
 ### Added
 - Parameter for stomp debug